Rename 'helloworld' to 'adr' and perform related refactorings
The commit includes renaming all occurrences of 'helloworld' to 'adr' in the opencdx-adr module. This includes changes to file and folder paths, package names, and application properties. The Dockerfile, docker-compose and Gradle build scripts were updated to reflect this change. Furthermore, updates are made in deploy.sh and in the SSL certificate generation script.