Refactored database table naming convention
Changed naming convention of tables in the database and affected codebase to use dimensions (prefixed with "dim") and facts (prefixed with "fact"). This change makes it easier to distinguish between different types of data, and enhances the organization and understandability of the model, following a more clear-cut method in line with star schema format.