Add OpenCDX processor implementations and refactor code
Introduced OpenCDX processors for Tinkar and Snowmed implementation. Refactored code to include new BaseAnfProcessor and applied it to existing processors. Moved method implementations in initial service class to the respective processor classes, promoting code reusability. The processor classes are now loaded dynamically using ApplicationContext and used in the OpenCDXAdrService, allowing for easier addition of new processor classes in the future.