Refactor priority and repetition models, update SQL schema
Refactored various models such as priority and repetition in the ADR module. Priority has been changed from an enumerated type to a regular String. The repetition model and the associated SQL schema has been updated to replace several integer and enumerated fields with references to the MeasureModel. This makes the code more modular, and easier to modify in the future.