Add LogicalExpressionUnitProcessor and unit reference in MeasureModel
A new LogicalExpressionUnitProcessor class has been created and referenced in the AppConfig. This processor enhances the ANF statement processing. Moreover, a new reference to the TinkarConceptModel has been added in the MeasureModel as 'unit' to align these entities with recent changes. Additional modifications have been made in existing classes and database scripts to accommodate these updates.