Update BeanShellSampler.query in OpenCDX.jmx
The code modifies the `BeanShellSampler.query` by adding a java code that generates a random date between January 1, 2019 and the present day. This date is then converted into epoch seconds and stored in a variable `anfSubmissionDate`. The `upperBound` and `lowerBound` values in various arguments are then updated to use this variable.